+++ NEWS.md
@@ -7,9 +7,13 @@ and their values in a nested list (used for dat-parsing).
 + `get_dat_param()` can be used to get values of a (valid) Zonation run 
 configuration parameter (i.e. the content of a dat-file) from a `Zvariant`
 object.
++ `set_dat_param()` can be used to set values of a (valid) Zonation run 
+configuration parameter (i.e. the content of a dat-file) for a `Zvariant`
+object.
 + `zparameters()` is used to fetch all accepted Zonation run configuration 
 paramteres (i.e. parameters in the dat file). Can return just the parameter 
 names or a list of `[[parameter_name]][[section]]`".
+
 
 ### BUG FIXES